requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20211102.0=hc377ac9_0
aiohttp=3.8.3=py39h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.5.0=py39hca03da5_0
appnope=0.1.2=py39hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h1a28f6b_0
arrow-cpp=8.0.0=py39h68e74d9_1
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py39hca03da5_0
attrs=22.1.0=py39hca03da5_0
aws-c-common=0.6.8=h1a28f6b_0
aws-c-event-stream=0.1.6=hc377ac9_5
aws-checksums=0.1.11=h1a28f6b_1
aws-sdk-cpp=1.8.185=h4a942e0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39hca03da5_0
black=22.6.0=py39hca03da5_0
blas=2.116=openblas
blas-devel=3.9.0=16_osxarm64_openblas
bleach=4.1.0=pyhd3eb1b0_0
blobfile=2.0.1=pypi_0
boost-cpp=1.73.0=h1a28f6b_12
bottleneck=1.3.5=py39heec5a64_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py39h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.18.1=h1a28f6b_0
ca-certificates=2023.01.10=hca03da5_0
certifi=2022.12.7=py39hca03da5_0
cffi=1.15.1=py39h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hca03da5_0
coloredlogs=15.0.1=py39hca03da5_1
comm=0.1.2=py39hca03da5_0
contourpy=1.0.5=py39h525c30c_0
cryptography=38.0.4=py39h834c97f_0
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.6.1=py39hca03da5_0
debugpy=1.5.1=py39hc377ac9_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.6=py39hca03da5_0
entrypoints=0.4=py39hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.9.0=py39hca03da5_0
flatbuffers=2.0.0=hc377ac9_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py39h80987f9_0
fsspec=2022.11.0=py39hca03da5_0
gettext=0.21.0=h826f4ad_0
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_1
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py39h8c48613_0
gnutls=3.6.15=h887c41c_0
grpc-cpp=1.46.1=h8e4afa7_1
huggingface_hub=0.10.1=py39hca03da5_0
humanfriendly=10.0=py39hca03da5_1
icu=68.1=hc377ac9_0
idna=3.4=py39hca03da5_0
ipykernel=6.19.2=py39h86d0a89_0
ipython=8.9.0=py39hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39hca03da5_1
jinja2=3.1.2=py39hca03da5_0
jpeg=9e=h1a28f6b_0
jsonschema=4.17.3=py39hca03da5_0
jupyter_client=7.4.9=py39hca03da5_0
jupyter_core=5.2.0=py39hca03da5_0
jupyter_server=1.23.4=py39hca03da5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=py39h313beb8_0
krb5=1.19.4=h8380606_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libblas=3.9.0=16_osxarm64_openblas
libboost=1.73.0=h49e8a49_12
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcblas=3.9.0=16_osxarm64_openblas
libcurl=7.87.0=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.8=h1a28f6b_5
libedit=3.1.20221030=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=hf27765b_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
liblapack=3.9.0=16_osxarm64_openblas
liblapacke=3.9.0=16_osxarm64_openblas
libnghttp2=1.46.0=h95c9599_0
libopenblas=0.3.21=openmp_hc731615_3
libopus=1.3=h1a28f6b_1
libpng=1.6.37=hb8d0fd4_0
libprotobuf=3.20.3=h514c7bf_0
libsodium=1.0.18=h1a28f6b_0
libssh2=1.10.0=hf27765b_0
libtasn1=4.16.0=h1a28f6b_0
libthrift=0.13.0=hd358383_6
libtiff=4.5.0=h313beb8_1
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=h68602c7_0
libwebp-base=1.2.4=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
llvm-openmp=14.0.6=hc6e5704_0
lxml=4.9.1=py39h2fae87d_0
lz4-c=1.9.4=h313beb8_0
markupsafe=2.1.1=py39h1a28f6b_0
matplotlib=3.6.2=py39hca03da5_0
matplotlib-base=3.6.2=py39h8bbb115_0
matplotlib-inline=0.1.6=py39hca03da5_0
mistune=0.8.4=py39h1a28f6b_1000
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.2.1=py39hca03da5_0
multidict=6.0.2=py39h1a28f6b_0
multiprocess=0.70.14=py39hca03da5_0
munkres=1.1.4=py_0
mypy_extensions=0.4.3=py39hca03da5_1
nbclassic=0.4.8=py39hca03da5_0
nbclient=0.5.13=py39hca03da5_0
nbconvert=6.5.4=py39hca03da5_0
nbformat=5.7.0=py39hca03da5_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=py39hca03da5_0
nettle=3.7.3=h84b5d62_1
notebook=6.5.2=py39hca03da5_0
notebook-shim=0.2.2=py39hca03da5_0
numexpr=2.8.4=py39h79ee842_0
numpy=1.23.5=py39h1398885_0
numpy-base=1.23.5=py39h90707a3_0
onnx=1.13.0=py39h8472c4a_0
onnxruntime=1.12.1=py39h7306aa3_0
onnxruntime-tools=1.7.0=pypi_0
openblas=0.3.21=openmp_hf78f355_3
openh264=1.8.0=h98b2900_0
openssl=1.1.1t=h1a28f6b_0
orc=1.7.4=hdca1487_1
packaging=22.0=py39hca03da5_0
pandas=1.5.2=py39h78102c4_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.10.3=py39hca03da5_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py39h313beb8_2
pip=22.3.1=py39hca03da5_0
platformdirs=2.5.2=py39hca03da5_0
prometheus_client=0.14.1=py39hca03da5_0
prompt-toolkit=3.0.36=py39hca03da5_0
protobuf=3.20.3=py39h313beb8_0
psutil=5.9.0=py39h1a28f6b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=8.0.0=pyhd3eb1b0_1
py3nvml=0.2.7=pypi_0
pyarrow=8.0.0=py39hd776c02_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.17=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39hca03da5_0
pyrsistent=0.18.0=py39h1a28f6b_0
pysocks=1.7.1=py39hca03da5_0
python=3.9.16=hc0d8a6c_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39hca03da5_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h1a28f6b_1
pytorch=2.0.0.dev20230216=py3.9_0
pytz=2022.7=py39hca03da5_0
pyyaml=6.0=py39h80987f9_1
pyzmq=23.2.0=py39hc377ac9_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2022.7.9=py39h1a28f6b_0
requests=2.28.1=py39hca03da5_0
responses=0.13.3=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.6.3=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=hc377ac9_0
sniffio=1.2.0=py39hca03da5_1
soupsieve=2.3.2.post1=py39hca03da5_0
sqlite=3.40.1=h7a7dc30_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py39hca03da5_0
terminado=0.17.1=py39hca03da5_0
tiktoken=0.2.0=pypi_0
tinycss2=1.2.1=py39hca03da5_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.11.4=py39h5c5695e_1
tomli=2.0.1=py39hca03da5_0
torchaudio=2.0.0.dev20230217=py39_cpu
torchvision=0.15.0.dev20230216=py39_cpu
tornado=6.2=py39h1a28f6b_0
tqdm=4.64.1=py39hca03da5_0
traitlets=5.7.1=py39hca03da5_0
transformers=4.24.0=py39hca03da5_0
typing-extensions=4.4.0=py39hca03da5_0
typing_extensions=4.4.0=py39hca03da5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py39hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39hca03da5_1
websocket-client=0.58.0=py39hca03da5_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xmltodict=0.13.0=pypi_0
xxhash=0.8.0=h1a28f6b_3
xz=5.2.10=h80987f9_1
yaml=0.2.5=h1a28f6b_0
yarl=1.8.1=py39h1a28f6b_0
zeromq=4.3.4=hc377ac9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.2=h8574219_0
